[fix] improve logic for determining finalization eligibility (#2476)
Refactors the logic for determining if a petition can be finalized.

Ensures the correct next incomplete field is selected when required
fields exist on the same page as the last completed field, even when
switching pages.

This addresses an issue where the next incomplete field was always the
first one in the page.

Also adds a guard to prevent errors when children fields are deleted.