[feat] show different ui for checkboxes when long list of options (#2489)
Enhances the display logic for checkbox fields in recipient view. It now
shows the multi-select component if a standard list exists or if the
number of values exceeds 15.

Also adjusts the styles for the multiselect to better align with the new
functionality, including hiding the input and adjusting the value
container's alignment.


![image](https://github.com/user-attachments/assets/b50b3bd2-82f0-4d3c-8e92-73ef85c084e4)


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209638888394120